requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.5
requests==2.18.3
RPi.GPIO==0.6.3
urllib3==1.22
